renamed helper functions to distilbert functions, added a starify() function.
